Update for Review class

+++ fpshoppingcart/src/main/java/com/lunifer/jo/fpshoppingcart/service/impl/ReviewServiceImpl.java
@@ -1,57 +1,49 @@
 package com.lunifer.jo.fpshoppingcart.service.impl;
 
 import com.lunifer.jo.fpshoppingcart.dto.ReviewDTO;
-import com.lunifer.jo.fpshoppingcart.entity.Product;
 import com.lunifer.jo.fpshoppingcart.entity.Review;
 import com.lunifer.jo.fpshoppingcart.mapper.ReviewMapper;
-import com.lunifer.jo.fpshoppingcart.repository.ProductRepository;
 import com.lunifer.jo.fpshoppingcart.repository.ReviewRepository;
 import com.lunifer.jo.fpshoppingcart.service.ReviewService;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.NoSuchElementException;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
+@Service
 public class ReviewServiceImpl implements ReviewService {
     private final ReviewRepository reviewRepository;
     private final ReviewMapper reviewMapper;
-    private final ProductRepository productRepository;
-
     @Autowired
-    public ReviewServiceImpl(ReviewRepository reviewRepository, ReviewMapper reviewMapper, ProductRepository productRepository) {
+    public ReviewServiceImpl(ReviewRepository reviewRepository, ReviewMapper reviewMapper) {
         this.reviewRepository = reviewRepository;
         this.reviewMapper = reviewMapper;
-        this.productRepository = productRepository;
     }
 
-    @Override
-    public ReviewDTO saveReview(ReviewDTO reviewDTO, long productId) {
 
+    @Override
+    public void saveReview(ReviewDTO reviewDTO) {
         Review reviewEntity = reviewMapper.reviewDTOToReviewEntity(reviewDTO);
-
-
-        Product product = productRepository.findById(productId).orElseThrow(() -> new NoSuchElementException("No se encontro el producto: " + productId));
-
-        reviewEntity = reviewRepository.save(reviewEntity);
-
-
-        return reviewMapper.reviewEntityToReviewDTO(reviewEntity);
+        reviewRepository.save(reviewEntity);
     }
 
     @Override
-    public void deleteReview(long reviewId, long productId) {
-        reviewRepository.deleteById(reviewId);
+    public void deleteReview(long reviewId) {
+        if (reviewRepository.existsById(reviewId)) {
+            reviewRepository.deleteById(reviewId);
+        } else {
+            throw new EntityNotFoundException("No se encontro review con ID " + reviewId);
+        }
+
     }
 
     @Override
     public List<ReviewDTO> getAllReviewsByProductId(long productId) {
-        List<Review> reviews = reviewRepository.findByProductId(productId);
-
-        return reviews.stream()
+        List<Review> reviewEntities = reviewRepository.findReviewsByProductId(productId);
+        return reviewEntities.stream()
                 .map(reviewMapper::reviewEntityToReviewDTO)
                 .collect(Collectors.toList());
     }
-
 }
